# Add a parameter to gfx_draw_glyph() for controlling blending behavior