This piece of code was added in a25f066 to eliminate text halo in the MangaGamer version of Sengoku Rance.

Text rendering is often a performance bottleneck. By omitting it for games other than Sengoku Rance, `gfx_draw_glyph()` becomes almost twice as fast.