@kichikuou 2024-06-13
While looking into https://github.com/kichikuou/xsystem4-android/issues/4#issuecomment-2161323954, I noticed that `SACT2.SP_TextDraw()` is a bottleneck in the equipment screen of Rance 6 (MangaGamer).

This change was originally made for the [Wasm port](https://github.com/kichikuou/xsystem4/commit/3302f885007fb66d0bc406a61d0ce37c323cf412), but I would like to merge it because it has benefits outside of Wasm too.