@nunuhara 2024-06-14
This doesn't seem to be working properly. See below comparison (top is current master, bottom is this patch).

![Screenshot_20240613_191608](https://github.com/nunuhara/xsystem4/assets/52369405/8def8442-ebc0-4538-91f7-93fc4b51d1e1)

I think what's happening here is that semi-transparent pixels from the outline (which is drawn first) are being colored black when the main glyph is drawn on top.

It seems to me that the solution would be to call `gfx_rance7_mg_prepare_draw_text` twice, once before drawing the outline, and then again before drawing the main glyph. So we would need a variant of `gfx_render_text` which makes these calls.