@nunuhara 2024-06-14
Actually, upon further testing, this issue is not specific to the MangaGamer version of Sengoku Rance. Rance Quest and Rance 01 also have this issue. For example, this is Rance 01:

![Screenshot_20240613_202642](https://github.com/nunuhara/xsystem4/assets/52369405/33ce9783-e004-47cf-9fa5-e751d1646a4b)

Other games I tested (Rance VI, GALZOO, JP Sengoku Rance) looked fine.

Notably, Sengoku Rance (MG version), Rance Quest and Rance 01 all use .fnl fonts. Perhaps `ft_font_get_glyph` and `fnl_font_get_glyph` are doing something slightly different which causes this issue only for .fnt fonts? Or perhaps older games are just coded in a way that doesn't trigger this problem.