@kichikuou 2024-06-14
Thanks for catching the problem! I overlooked the case of edged fonts.

When I tested the Wasm port, I tried every Japanese game marked as supported in game_compatibility.md, so it's probably a problem specific to SengokuRanceFont and PartsEngine.

It would be ideal if we could deal with this by changing the texture initialization / clear color. I will check the code of MangaGamer Sengoku Rance and Rance Quest to see if that works.