@kichikuou 2024-06-15
I was thinking about initializing a texture with the text color, but that doesn't seem very robust. For example, it doesn't work well with the following sequence:

```
sact_text_metrics_t tm;
tm.nColorR = 255;
SP_TEXT_DRAW(1, "foo", tm);
tm.nColorR = 0;
SP_TEXT_DRAW(1, "bar", tm);
```

So I decided to change my approach and add an argument to `gfx_draw_glyph()` to control the behavior of blending. When drawing on a transparent texture (like a text sprite), I don't think it is a good idea to use alpha blending. The text color should be written as is, and only the alpha value should be variable.

I'm still testing it out, but please take a look and let me know what you think.