@nunuhara 2024-06-16
I'm think this is a good approach. I only tested briefly but it seems to be working. Let me know when you're done testing and I'll merge it.