@kichikuou 2024-06-17
I was testing it on Android and noticed that text is sometimes chipped.

![Screenshot_20240618-082612](https://github.com/nunuhara/xsystem4/assets/12679772/77c21795-f9bc-4f35-9bbc-e61503344aeb)

This is very strange:
- It only happens on one of four Android devices I've tested (Fire HD 8, 10th gen). It does not reproduce on desktop.
- If I re-add the following code in `gfx_draw_glyph()`, which I removed in this PR, it does not reproduce.

```c
	glBlendFuncSeparate(GL_ONE, GL_ZERO, GL_ZERO, GL_ONE);
	struct copy_data fill_data = COPY_DATA(dx, dy, dx, dy, glyph->w * scale_x, glyph->h);
	fill_data.r = color.r / 255.0;
	fill_data.g = color.g / 255.0;
	fill_data.b = color.b / 255.0;
	fill_data.a = 0.0;
	fill_data.threshold = 0.001;
	run_copy_shader(&fill_amap_under_border_shader.s, dst, dst, &fill_data);
```

- Just removing the above code from the current master produces the same symptoms.

I suspect this is a bug in the OpenGL ES driver of the device, and at least not due to the code added in this PR, but let's not merge this until I find a workaround.