@kichikuou 2024-06-18
Pushed an additional commit that makes `gfx_init_texture_rgba()` use `glClear()`. This avoids the Fire HD 8 issue mentioned above (I don't know why) and also has some performance benefits.

I ran a benchmark with the following code:

```c
   uint64_t start = SDL_GetPerformanceCounter();
   for (int i = 0; i < 10000; i++) {
       struct texture t;
       gfx_init_texture_rgba(&t, w, h, (SDL_Color){0, 0, 0, 0});
       gfx_delete_texture(&t);
   }
   uint64_t end = SDL_GetPerformanceCounter();
   WARNING("Texture creation/deletion time: %f ms", (end - start) * 1000.0 / SDL_GetPerformanceFrequency() / 10000);
```

Here's the results on Fire HD 8, 10th gen:

| Size | Current master | This PR |
| ---- | -------------- | ------- |
| 128x128 | 0.365 ms | 0.363 ms |
| 256x256 | 0.553 ms | 0.374 ms |

Now I think this PR is good to merge.